+++ module-admin/src/main/java/com/adminapplication/login/WebMvcConfig.java
@@ -1,4 +1,4 @@
-package com.adminapplication.admin;
+package com.adminapplication.login;
 
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
